Cleanup old run views.
