--- webapp/src/components/backstage/playbook_runs/playbook_run_list/seach_input.scss
@@ -1,37 +1,0 @@
-.PlaybookRunList-search {
-    position: relative;
-    max-width: 56rem;
-    width: 100%;
-
-    input {
-        -webkit-transition: all 0.15s ease;
-        -webkit-transition-delay: 0s;
-        -moz-transition: all 0.15s ease;
-        -o-transition: all 0.15s ease;
-        transition: all 0.15s ease;
-        background-color: transparent;
-        border-radius: 4px;
-        border: 1px solid var(--center-channel-color-16);
-        width: 100%;
-        height: 4rem;
-        font-size: 14px;
-        padding-left: 4rem;
-
-        &:focus {
-            box-shadow: inset 0 0 0px 1px var(--button-bg);
-            border-color: var(--button-bg);
-        }
-    }
-
-    &:before {
-        left: 18px;
-        top: 9px;
-        position: absolute;
-        color: var(--center-channel-color-56);
-        content: '\f349';
-        font-size: 20px;
-        font-family: 'compass-icons', mattermosticons;
-        -webkit-font-smoothing: antialiased;
-        -moz-osx-font-smoothing: grayscale;
-    }
-}
