+++ webapp/src/components/backstage/runs_list/status_filter.tsx
@@ -3,9 +3,8 @@
 
 import React, {useState} from 'react';
 
-import styled, {css} from 'styled-components';
-
 import {MultiSelect, SelectOption} from 'src/components/multi_select';
+import {PlaybookRunFilterButton} from '../styles';
 
 interface Props {
     default: string[] | undefined;
@@ -17,52 +16,6 @@ export interface StatusOption {
     value: string;
     label: string;
 }
-
-interface PlaybookRunFilterButtonProps {
-    active?: boolean;
-}
-
-export const PlaybookRunFilterButton = styled.button<PlaybookRunFilterButtonProps>`
-    display: flex;
-    align-items: center;
-    border: none;
-    padding: 8px;
-    border-radius: 4px;
-    color: var(--center-channel-color-56);
-    cursor: pointer;
-    font-weight: 600;
-    font-size: 14px;
-    line-height: 12px;
-    -webkit-transition: all 0.15s ease;
-    -webkit-transition-delay: 0s;
-    -moz-transition: all 0.15s ease;
-    -o-transition: all 0.15s ease;
-    transition: all 0.15s ease;
-    padding: 0 16px;
-    height: 4rem;
-
-    :hover {
-        background: var(--center-channel-color-08);
-        color: var(--center-channel-color-72);
-    }
-
-    :active {
-        background: var(--button-bg-08);
-        color: var(--button-bg);
-    }
-
-    .icon-chevron-down {
-        :before {
-            margin: 0;
-        }
-    }
-
-    ${(props) => props.active && css`
-        cursor: pointer;
-        background: var(--button-bg-08);
-        color: var(--button-bg);
-    `}
-`;
 
 export function StatusFilter(props: Props) {
     const opts = props.options.map((opt) => ({
